Resolves #109: add new RecordCursor API to make correct use easier.
This change introduces a new API for advancing a cursor using the
onNext() method, which returns future for a structure that is either:
- a next result and a non-end continuation, or
- a continuation and a reason that there was no next result.
This change provides a reimplementation of all existing cursor to use the
new API. All cursors also support the old Iterator-style API using shims
that have as little logic as possible. The idea is that we can
eventually remove these without needing to change much code.

In implementing this change, a dozen (or two) small bugs were fixed,
mostly involving strange edge cases of continuations and
no-next-reasons. In a few cases, this required changing tests that relied
on buggy or out-of-contract behaviour.

Additionally, the very beginning of #110 was implemented. Since the new
API materializes a continuation on every call to onNext() (rather than
on request), some rudimentary structured continuations were added to
reduce the amount of time spent serializing to byte arrays.

With this change, the only code that uses the new API are cursors
themselves.