Add phasing to rewrite planner and allow rules to report changes made.
This includes two vaguely related changes that are needed to make the
planner actually work:
- making the rewrite planner work with multiple phases of rules,
where possible rule applications in a phase are exhausted before trying
to apply any rules in the second phase. This turns out to be important
as we port planning rules over from the old planner.
- allowing rules to report whether or not they made changes. This makes
it easy to tell which rules require us to "recheck" for more rule
applications, and which rules simply failed to do any useful work after
matching. It's also an easy way to avoid certain infinite loops in
planning logic.

Along with those changes, this commit fixes a bunch of bugs in the
rewrite planner.