Add a property visitor for the set of returned record types.
Includes an implementation of a property (as a visitor) that returns the
set of record types that a planner expression might return. Also inclues
some small refactoring to make this easier/more robust to write, by
creating interfaces for relational planner expressions to expose the
number of relational children that they have and for returning the
record type filters that they include.