add FilterSatisfiedMask class to keep track of which filters have been used
In theory, this would allow the planner to keep track of what filters have already been used across different invocations, which can be useful in doing things in more than one pass. It also allows for a more nuanced notion of how a filter may or may not be satisfied, handling the tree-like nature of a query component better. This class is currently unused (not counting tests).