Resolves #24: Full Text: Add support for querying the values of maps
This involved changing the QueryToKeyMatcher to support matching only up through the end of a key and being (therefore) okay with missing some filters. This begins the work necessary to complete #21, though it doesn't finish it. The QueryToKeyMatcher then matches the grouping part of the key expression, and some logic within TextScan matches the text part and produces a text query.

This also introduces #53. There is some work in the FilterSatisfiedMask class that should be able to fix this if done correctly, but unfortunately, that hasn't been accomplished yet. I figured it would be better to make the planner good enough to handle maps and then work out how to have it not handle malformed map queries later.