Resolves #25: Support covered queries with text index scans
This refactors the RecordQueryCoveringIndexPlan class so that it now has a child plan. That plan has to be able to produce a stream of index entries, and then the covering plan turns those streams of entries into partial records (as was done before this change). The reqular and text index scan plans have then been made to comply with this interface so that they can be used with the covering scan. In addition, to support using covering indexes with prefix queries, this also changes the logic to push down a covering query so that the covering operation is applied before applying the unordered primary key distinct filter.