Resolves #73: Produce artifacts with shaded protobuf and guava dependencies
This adds a new subproject whose sole purpose in life is to produce the shaded artifacts. It takes the core project as a dependency, and then it produces a shadow jar containing that dependency, guava, and protobuf-java. The core project is kept in its namespace; the two others are now under the com.apple.foundationdb.record.shaded package.

Left as an excercise to the reader is how to produce output from protoc whose generated message are in our shaded package path rather than in the standard package.