Fixes #117: Full Text: Queries on maps when combined with other filters throw an error when planned
The fix was to make the QueryToKeyMatcher more recursive and have it no longer try and guess which path it needs based on a string representation of how the path from the root to the leaf element would look. This allows it to handle nested ands, etc. Depending on what kinds of queries and expressions it gets, it might be less efficient in some cases if it needs to do a lot of recursing.