Resolves #149: Add a mechanism for determining statistics about records/indexes
The SizeStatisticsCollector is that mechanism. The class allows the user to specify a subspace, either directly or by specifying an index or a store they want statistics on. It then allows them to call "collect" with an FDBRecordContext. That takes an ExecuteProperties which can be used to specify some kind of limit to the request. It handles continuation handling for the user, and if it encounters a retriable error, it will stop executing, but it will not throw an error. (That is, it will behave as if the limit is reached from the outside point of view of the user.) This allows the operation to be retried with a new transaction without hitting the same key twice.