Resolves #196: Add a method on FDBRecordStore to check if a record exists
The method is called "recordExists", and there are blocking and asynchronous variants of each. There are also variants that take an IsolationLevel and variants that don't. (By default, it reads at SERIALIZABLE isolation level.) At present, this method just runs "loadRawRecordAsync" in the background and then returns a result (without serializing the record or adding it to the preloadCache, though records will still be loaded into the read-your-writes cache). Future work could make use of a future (but unimplemented) FoundationDB client API that allowed the user to ask if a key existed without returning the values (to save network traffic).

This also resolves #194 as it makes the loadRawRecordAsync function private, so now FDBRawRecord and loadRawRecordAsync have sensible visibility restrictions when paired together. In my mind, the most sensible reasons for an external user to want to use loadRawRecordAsync is to either (1) get the side-effects of loading the record into the cache (which can be done by calling preloadRecordAsync) or (2) check if the record is present or not by comparing to null (which can now be done by calling the new method).