Resolves #3: Use evolved meta-data file descriptor to build meta-data
Storing a record to a record store used to fail, if the record's meta-data was an evolved version of
the record store's meta-data shared using a meta-data store. This change should fix the issue by
passing the evolved meta-data to the builder.

Notes:
- The record types are now extracted from the union descriptor.
- If local meta-data is provided, the original union descriptor is used
as a base but the descriptors are taken from the local meta-data file.
- Added local file descriptor to meta-data store and dropped it from record store.
- Also added validation for both records and local descriptors. It's
happens both in setRecords and in MetaDataValidator.
- Modified FDBRecordStoreTest.open test to set the extension registries
of meta-data store. It looks like proto3 requires it.
- Modified some tests to use a new meta-data store for saving the record.