Fixes #222: The FlatMapPipelinedCursor can return null continuations when not done
This now remembers instances where the cursor hit a limit that made it stop and saves the continuation state from that run. This is then used to calculate the continuation in situations where the cursor stopped before returning any results.

This also (possibly) changes the contract if the inner cursor stops with an in-band limit other than SOURCE_EXHAUSTED (e.g., return limit reached). The previous behavior was to consider that inner cursor done and move on to the next one. This was (arguably) a bug in that it could lose results, but on the other hand, if the user just wanted the first result from each inner cursor, then that was doing the right thing. That particular part of the PR could be reverted and the test can be modified to change the semantics of what that means.