Fixes #227: The OnlineIndexerTest::closeWhileBuilding test sometimes hangs
There was an instance where we return a future and update it within a whileTrue loop. If the runner was closed at the right time, then it could be the case that the whileTrue loop itself ended in an exceptional state. That error was swallowed and the returned future was never completed. This adds a check to that whileTrue statement to catch such an error and propagates it to the returned future.

This also adds a 5 minute timeout to operations in the OnlineIndexerTest so that errors like this eventually cause timeouts rather than things infinintely hanging.