Fixes #241: Intersection cursors can return null continuations when children hit certain in-band limits
This was caused if one of the cursors stopped for an in-band reason and another cursor happened to have a null continuation (which can happen if the intersection cursor hasn't consumed that cursor yet). The replacement is a more direct test to see if any child stopped because it was exhausted as well as some comments as to what it's supposed to be doing.