Resolves #317: Make the PlannerProperty visitor more expressive.
Prior to this change, writing properties as visitors was very tedious.
Most properties can be expressed as a depth-one recursion, so augmenting
the visitLeave() methods with the results of calling visitLeave() on its
children (or members, in the case of references) makes properties easier
to read and write. This change also renames the visitor methods,
changing visitEnter() to shouldVisit() and visitLeave() to
evaluateAtExpression() and evaluateAtRef(), which also aids in
comprehension.