Fixes #255: FlatMapPipelinedCursor can return a null continuation when outer cursor is not exhausted if the first future is delayed
This makes it so that the only way for "tryToFillPipeline" to return "false" is if either (1) the first entry in the pipeline is the "outer cursor done" sentinel value or (2) the inner cursor of the first item of the pipeline has a new item. Otherwise, it will always loop, so it will always add the sentinel upon outer cursor completion. Or that's at least the idea.