Resolves #262: Detect blocking calls in an asynchronous context
The purpose of this change is to attempt to detect when a blocking call
(`FDBDatabase.asyncToSync()`, `FDBRecordContext.asyncToSync()`,
`CompletableFuture.join()` and `CompletableFuture.get()`) are made while in
an asynchronous context; specificaly those that take place within a
`CompletableFuture` completion stage.  This behavior is undesirable because
it can lead to exhaustion of the executor thread pool and an effective
deadlock situation (ask me how I know).

This detection is performed by grabbing a stack trace just prior to the
blocking operation and seeing if the current thread's stack involves any
`CompletableFuture` method. This approach is pretty crude, but seems to be
effective.

There is also one additional scenario that is watched for in this logic
and that is an attempt to perform a blocking call in a method that is
producing a future.  By contention, the Record Layer library ends methods
returning futures with `Async`, so we treat any method named as such as
a possible source of this problem. There are some legitimate reasons in
which this may be taking place, so these are always logged as a warning.

The code changes introduced are as follows:

- `FDBDatabaseFactory.setBlockingInAsyncDetection()` allowing you to enable
  logic in asyncToSync calls that checks to see if they are being made from
  within a future completion stage.
- `BlockingInAsyncDetection` describes how the code should react to the
  situation in which asyncToSync is being called from within a future
  completion.
- `CKFDBDatabase.asyncToSync()` updated to perform the detection
- Added `FDBDatabase.get()`, `FDBDatabase.join()`, `FDBRecordContext.get()`,
  and `FDBRecordContext.join()` that will perform the same sort of detection
  as `asyncToSync`. I have selectively updated a few places to use these
  methods, but I am sure there are more locations in which this can be done.
- Created new abstract `FDBTestBase` class that checks a system property
  (`com.apple.foundationdb.record.blockingInAsyncDetection`) to configure whether or
  not detection is enabled and how to respond to the presence of blocking in async.
- Added testing infrastructure to create assertions on logging messages.
- Updated gradle to enable async detection in our unit tests. I did some benchmarking
  and found that the detection did not noticably impact the performance of the
  tests.