Address #328: Generate sources and javadoc jars for the shaded artifacts
This only "addresses" #328 rather than "resolves" it as this does not correctly merge the sources or fix the shaded paths. It does, however, generate non-empty sources and Javadoc jars for the shaded artifacts, which is good enough for some amount of work. In particular, I was able to verify that IntelliJ was able to use these Jars and give Javadocs and fill in parameter names, which is a good start. I think I would still prefer if it combined the Jars correctly, but this probably is good enough to get most people unstuck.