Fixes #332: UnorderedUnionCursor can sometimes change its mind about whether there are more elements
This modifies the behavior of an UnorderedUnionCursor to keep returning results as long as there is at least one child cursor with more to say. The problem was that an intervening completion of a sibling cursor could turn a cursor from something that had next to something that does not. This removes the bug by making "anyHaveNext" just, like, check to see if any have next.