Resolves #336: Add an unordered variant of the IntersectionCursor
This does not use the unordered intersection cursor in any way, but it adds it to make future work possible. It comes with a few caveats in order to make it possible to resume from a continuation. In particular, it can have false positives, so any plan that uses it either needs to be okay with that or it needs to check the results of the intersection (if possible) to filter them out. But it is guaranteed to return everything that really is in the intersection, so no results are missed.