Resolves #343: Add text predicates for "all prefixes" and "any prefix"
This exposes them in a way that is congruent with the other text predicates. It also allows the user to tweak properties of the index scan through parameters of the query predicate because that was the easiest place for me to put it, but I am not sure it is right.