Resolves #349: Add limit for number of bytes scanned by a cursor.
This adds a new limit ("the byte scan limit") that's broadly similar to
the record scan limit but that counts the number of bytes being scanned
instead of the number of key-value pairs. The semantics of the
implementation are pretty weak because we can't in general know the
number of bytes that a cursor will scan when it asks for permission to
continue. However, the implemented limit should still provide some kind
of protection against queries that scan a lot of big records.