Fix #251:  Directory layer cache size was accidentally defaulted to zero
Tests were showing an excessive number of read-only commits taking place.
This was due to the fact that the directory layer cache was accidentally
set to a size of zero by default, meaning that every request to look up
an entry required a new transaction.

As a side note: we could buy some efficiencies by doing the initial read
within the existing client transaction and only use a secondary transaction
in the case that the entry needs to be created. I think that this path wasn't
taken in the current code because it makes the semantics of passing a
context into the resolve() call a little confusing.