Resolves #354: Test versionstampIndexDeserialization seems broken
This fixes the test which was mentioned in the issue definition, and it adds the validation logic that makes sure that there is at least one record type defined in the meta-data. There were a few tests where the meta-data was being created without any records, so those have also been changed to not do that. The validation check should be harmless for existing record stores in that anyone who has successfully inserted data must have created at least one record type. At the moment, the primary cause of creating a RecordMetaData without any record types is failing to define a primary key for a record. This should no longer be the case after #3 is resolved as that will have its own error.