Improve DualPlannerExtension to be compatible with ParameterizedTest.
Previously, the DualPlannerExtension was incompatible with the
ParameterizedTestExtension which is used in many planner tests. By
default, JUnit will find all extensions of this type and run each of
them in sequence rather than taking the Cartesian product. As a result,
the tests runs with the DualPlannerExtension have unbound parameters.
The new implementation of the DualPlannerExtension identifies whether
the ParameterizedTestExtension is also registered for the same test; if
it is, it "nests" them appropriately. This is a bit of a hack but it
works well enough.