Resolves #56: Add rules for planning sort-only queries as index scans.
Adds a variety of rules for planning sorts as index scans. With some
minor modifications, this also supports pushing down filters that can be
satisfied by the same index to the index scan level using the methods in
ScanComparisons. As a result, a sizeable number of our current planning
tests now pass against the experimental planner. Also fix some bugs in
the implementation of phasing in the RewritePlanner.