Fixes #358: ChainedCursor does not obey byte/record/time scan limits (#371)
Fixes #358: ChainedCursor does not obey byte/record/time scan limits

The `ChainedCursor` was relyig on any cursor that it was wrapping to enfoce
scan limit (that is, if it was even wrapping a cursor at all).  The problem
was that in almost any sane use case, the cursor that it would be wrapping
is going to be asked to read only signle row at a time (see
`KeySpaceDirectory.listAsync` for an example of this), and our cursors have
the policy that they will *always* be able to make some progress.  That is,
they will allways be able to advance one element, regardless of whether
our limits have been reached.  Due to this behavior, the `ChainedCursor`
would never actually stop.  Aditionally, even if it was not wrappping
a cursor that did reads, it would never obey the time limits.

This change promotes `ChainedCursor` up to a `BaseCursor` in that it will
now do the policy enforcement of limits. It retains a way to build it in
which it will not do any enforcement (other than returned rows) if needed.