Resolves #356: Add limit on the size of the DNF during normalization.
Currently, the RecordQueryPlanner tries to normalize Boolean predicates
into disjunctive normal form. The DNF of some expressions can be
exponentially larger than the given form; in the worst case, this can
cause out-of-memory errors. This adds a limit to the Boolean normalizer
so that it will not attempt to normalize expressions with a DNF size
larger than the limit, using methods which determine the size of the DNF
without computing it.