Resolves #419: Rework ExpressionMatchers for complex child matching.
Adds a new ExpressionChildrenMatcher interface that does the work of
applying some matchers to the children of a planner expression, and
several implementors. Reworks the matching structure and pushes more
work (such as producing bindings) into the ExpressionMatcher and related
classes while simplifying the Bindable interface. Makes most types
related to matching, including PlannerBindings, immutable to sidestep
the tricky issue of incompatible bindings. Adds support for a single
match to produce multiple bindings and combine them appropriately (using
a Cartesian product if necessary).