Resolves #366: Add rules for planning on multi-field indexes.
Adds rules to plan a number of filters as a scan of a single multi-field
index. To support this, several new pieces of infrastructure were
needed:
1) A new PlannerExpression to allow a very specific type of "group
expression" within the semantics of the existing RewritePlanner and a
set of rules to manipulate such an expression.
2) A new matcher to match one of the children of a planner expression to
a particular matcher while providing a reference to the rest of the
children.
3) A data structure for tracking which part of a KeyExpression have been
matched with a comparison and a new class for expression a range of
comparisons.
This commit also includes some large changes to existing rules and
several new rules for implementing the new planning behaviour.