Fix #458: KeyWithValueExpression does not split correctly with FanOut.
The getKeyExpression() method on KeyWithValueExpression is supposed to
produce the "key" side of the KeyWithValueExpression. The current
implementation uses normalizeKeyForPositions(), which is incorrect in the
case of a FanOut. It now uses getSubKey(), which is correct.

This required some downstream reaction to properly handle
FunctionKeyExpressions, which are a bit special.