Resolves #441: Audit code for incorrect uses of CompletableFuture::isDone
Nothing major was found. There were a few instances where we were not using the cached READY_TRUE and READY_FALSE futures, so I fixed those. A few places, the error was generated and then caught, so I instead made it check for normal completion so that propagating the error was a little cheaper. I did not touch the MapPipelinedCursor or FlatMapPipelined cursor (except for a few comments), though I tried to add tests to excercise some of the code paths that might encounter exceptions. Those were all either directly or later calling .join() on the future they were checking, so they should all report errors, though there may be ways to fix those up so that they do not need to throw. I started down that path, then I decided that it was probably better to avoid accidentally introducing a new bug under the axiom of not trying to fix what is not Baroque.