Resolves #204: Add a default union descriptor when FDBMetaDataStore is used
This change adds a missing union descriptor to the record meta-data if FDBMetaDataStore is
used. The assumption is the following steps are involved:

- A .proto file is created that has some record types but doesn't have a union.
- The .proto file is saved using FDBMetaDataStore.saveRecordMetaData(). The missing union
 is added using saveRecordMetaData before it saves the meta-data into the store.
- Users can continue adding record types to this meta-data, which will add a field to the
now-existing union.

It is important that we support union-less local file descriptors too.
Otherwise it will be even harder for users after this change, because
they will need to deal with dynamic messages for storing records.

To solve this problem, we add a synthetic union based on the union
that's stored in the meta-data store.