Fixes #494: Preload cache is not invalidated after a record write
This invalidates the cache after the old record has been loaded but prior to when the new record is about to be written. This means that subsequent readers will then see the written value grâce à the RYW cache. There remains a few edge cases related doing things like, say, concurrently modifying a record while reading it. This does not make that problem worse, though, so maybe a more thorough discussion as to what needs to happen is required.