Fixes #526: ProbableIntersection and UnorderedUnion cursors can become stuck if one child completes with a limit and another an error
This makes it so that whenAny only looks at cursors that might return a next value, so if the cursor has hit a limit, it is now excluded from the list of cursors to investigate. This actually can also cause an infinite loop if there is a cursor that hit a limit and another cursor who has not yet completed in which case it would spin loop waiting for that cursor to complete. Note that this could end up blocking the thread that asked for the future if the other future completed instantaneously.