Resolves #524: Invalidation callbacks in the AsyncLoadingCache can build up
This makes the async loading cache only add the call back if necessary. In particular, it eschews adding the call back if the future is already in the cache or if the supplied future is already completed without error. This still allows for the same callback to be added multiple times to the same future (in particular, if the same key is accessed concurrently from a blank cache, then they both might add the callback), but it will add the call back in many fewer places. Hopefully, this reduces contention around the cache key and also reduces the number of these classes that are chained up together.