Resolves #410: Cache store header and index state
This adds a cache for `RecordStoreState`s that is keyed by the read version of the transaction. (This is the best we can do before 6.1 without introducing something like change locking.) This:

* Adds the store header to the `RecordStoreState` class. This makes sense as the `RecordStoreState` is supposed to represent all meta-data state that is specific to a record store. (In other words, it distinguishes record stores that share meta-data.) This seemed to be something that made sense packaging together with the index meta-data.
* Adds the FDBRecordStoreStateCache interface and two implementations. The cache is database-scoped and validates that it is not used against the wrong database. The first implementation always talks to the database and essentially is present to avoid some null checks in the record store implementation. The read version implementation caches record store states based on a pair of subspace provider and read version. Then the record store will add the cache entry to the cache when it is loaded from the database.

The current implementation doesn't buy the user *that* much, and it can use a non-trivial amount of memory storing cache entries. (Though one can set a maximum size to help combat that.) The main thing is that it lays the ground work for a better cache based on the global meta-data version key that was added in 6.1.