Resolves #538: AsyncLoadingCache should cache values instead of futures
The `AsyncLoadingCache` is now backed by a `Cache<K, V>` instead of a `Cache<K, CompletableFuture<V>>`, and if the value is not cached, then each caller creates their own future. This produces duplicate work, but it also creates more failure indepdendence (hopefully). The first future to complete is the one whose value is cached, and all completing futures get that value.

Some of the discussion related to this change also suggested removing the deadline. I elected to keep it (in the name of preserving semantics), but it is also now explicitly the case that the caller can provide a deadline of `Long.MAX_VALUE` to indicate that there should be no deadline imposed on the futures.