Resolves #506:Index meta-data
This change adds a mechanism for storing index meta-data. The index
meta-data is stored in a different subspace than where index state is
stored. The reason is, in contrast to index state, index meta-data
includes READABLE indexes too, which under some circumstances it could
add to the overhead of querying non-readable indexes.

Resolves #561 too.

Notes:
- This PR modifies some of the tests to add some coverage. But I think
it could benefit from having more tests.
- There is one question for the reviewers in a TODO. While loading the
index meta-data, we could discover some indexes that do not exist in the
record meta-data. In those cases, I'm skipping them. Let me know if you
think this is suboptimal.