Resolves #561: Allow storing READABLE indexes in the record store state
This change modifies the existing behavior that marks an index readable
by removing it from the state. The idea here is, an index is now
readable either if it's not present in the record store state, or it's
present with `state = READABLE`. The first `markIndexReadable` will put
the `READABLE` index in the map.

This will become more interesting after issue #506 is done.