Fixes #586: Several ProbableIntersectionCursor tests require FDB but are not marked as such
This was somewhat subtle, but there were some tests of the probable intersection cursor that relied on `AsyncUtil.whileTrue(Supplier)`. That will then run its tasks on the default FDB executor (instead of the default Record Layer executor or something like that). That executor, `FDB.DEFAULT_EXECUTOR`, requires the `FDB` class to be initialized, which because of reasons, is the thing that actually loads the native library. So if the native library is not present, then that class initialization fails.

The reason the `prb-sonar` task is passing is that this is only problematic in places where the native library is not installed. The `prb-sonar` task runs within an environment where the native library is present, but the server is not started. Running this test in a place with neither the server nor the client produces the bug.