Resolves #588: Repair records missing unsplit record suffix
This adds a new method to FDBRecordStore to allow you to iteratively
repair a record store that is in the following state:

- The store does not allow split records
- The format version indicates unsplit records must have a suffix indicating as such
- There are records that are missing this suffix

The logic will identify those records missing the suffix and attempt to
tack it back on.

As a note: this repair logic can only work with primary key expressions
in which the record key is always a fixed length. As such, it adds a method
to `KeyExpression` to allow you ask the question "are you fixed length"?
Many of the changes in this PR are to add this method to all of the expressions.