Remove ThenKeyExpression column caching
I had added caching so that `ThenKeyExpression.getColumnSize()`
could compute the number of columns only once, but this started
causing random test failures...and I do mean random, for some
harnesses the tests might pass on one run, a few fail the second
run, an entirely different one fails after that.  I don't really
have an idea as to why this is happening, but this change wasn't
really needed for this issue anyway, so I backed it out.