Resolves #508: Allow for record type renames in the evolution validator
This computes a mapping of renamed types based on using their positions in the union descriptor as the source of equivalency. (Nested types do not need to be checked.) It also validates that all of the usages of the top-level record type has been changed in the list of record types as well as for any index definition. The `MetaDataProtoEditor` has also been updated to include a method that will update the record types in `MetaData` proto message. It updates both usage of the record type within the record definition as well as in record type and index definitions.

There is a fair amount of complexity stemming from the fact that Protobuf employs a fairly complicated scheme for resolving packages and scopes. The correct way to do this, in some sense, is to create a tree based on the different types that are defined, etc., but I didn't do that. What I did, was for fully-qualified field types (which can be distinguished because they begin with a "."), the field type can be determined by basic string comparison. For non-fully qualified field types, one can determine if the field might be the right type by basicaly seeing if there the field matches the type we're looking for when prepended by some of the scopes from the outer name space. In that case, the best we can do is say, "this *might* match", and we can throw an error to avoid corrupting the record definition further. Protoc-generated records definitions will use fully-qualified field types for their generated fields, so that's good, but I'm still not super happy with how that turned out, if I'm honest.

Fixes #620: Union fields are assigned based on declaration order rather than field number.

If multiple union fields appear of the same type, the last one (by field number) is preferred instead of by declaration order.