Resolves #551: Increase required FDB version to 6.1
This is a fairly straightforward change. This updates the version numbers of both our FDB Java dependency and the server versions used within our build docker images. It also updates the API version, which isn't strictly necessary (as we were not using any of the things that changed with the API version change), but we will soon be relying on 6.1 features, and this stops someone from accidentally pointing a Record Layer library at a 6.0 FDB cluster.