Resolves #537: Invalidate store state caches using the meta-data version key in FDB 6.1
This adds a new implementation of the FDBRecordStoreStateCache that caches by the meta-data version-stamp key made available in FDB 6.1. Actually using the new cache requires both setting the database to use the new cache as well as marking any record store whose state should be cached as "cacheable" in its store state.