Introduce IndexEntrySource abstraction to represent general scans.
Currently, there are two different ways to get a stream of record
primary keys: a scan of a range of an index, and a scan of the
records by primary key. There are important distinctions between
the two during query execution because the former requires looking
up the records by primary key. However, the two operations are
basically identical for most planning tasks. This commit creates a
unified abstraction for the source of general index entries in the
form of the IndexEntrySource and the corresponding
IndexEntrySourceScanExpression. This abstraction substantially
simplifies several key planning operations, including predicate and
sort implementation.