Resolves #367: Add planner rules to support filters on nested fields.
Introduce the "NestedContext" as an abstraction to represent that an
entire planner expression sub-tree should be viewed relative to some
particular path through the nested records. Add methods to various types
that appear in the intermediate representation to "nest" and "unnest"
with respect to a given nested context.

Currently, this only implements filters on non-repeated nested fields,
but the infrastructure for oneOfThem()-type filters on repeated fields
is there.