Resolves #359 and #501: Plan OR predicates with an unordered union (#503)
* Resolves #359: Plan OR predicates with an unordered union

This introduces a new logical unordered union plan and two rules, one which turns a filter over an OR predicate into a logical union and another rule which implements a logical unordered union into the appropriate record query plan.

* Resolves #501: Planner should filter out duplicates with a PrimaryKeyUnorderedDistinctPlan

This adds a logical "distinct" plan and a default implementation thereof that will add the unordered primary key distinct plan. It will also only do so if the plan that was produced for its child produces duplicates. At the moment, the only way to make this happen is with an UnorderedUnion as the new planner does not support repeated fields which is the other way to get a duplicate record back.