Resolves #648: Add support for planning OneOfThemWithComponent.
Add support for planning predicates on nested records in a repeated
field against nested indexes on fields with a FanType of FanOut. This
uses the same NestedContext mechanism as non-repeated nested planning,
but uses somewhat different code paths for manipulating the
KeyExpressionComparisons. Also, simplify the NestedContext system to use
a single class with a boolean method, instead of an interface with
multiple implementations.