Resolves #584: Expose trace log format option through FDBDatabaseFactory
The `FDBTraceFormat` enum can be used to specify whether the trace logs should be emitted in XML or JSON format. It allows includes a "default" option that follows the system default. This means this change backwards compatible with anyone who has used the `FDB_NETWORK_OPTION_TRACE_FORMAT` environment variable to do the same thing. If both the environment variable and a (non-`DEFAULT`) log format are set through the new interface, the new interface wins.