Resolves #650: Create no-op operation to test network thread liveness/responsiveness
This adds the `performNoOp` and its async variants to the `FDBDatabase` class. This also adds events and waits for that which can be instrumented.

The exact no-op thing that is done is to create a new transaction, set its read version, and then attempt to read it. That requires performing no I/O (as verified by trying to run against a fake cluster). It is also harder to test, but if one calls `stopNetwork`, this will catch the fact that the network thread is down. I have not been able to reproduce the network thread just being really, really slow, but alas.