Resolves #514: Store index state with subspace key
The index state used to live under the
common_record_store_prefix + (5L, the name of the index). The intent of
this commit is to store the index state under the index subspace key
instead of the index name. In favor of clarity during upgrade, this
change also adds a new key space for the new indexes states (9L). Adding a
key space is not strictly necessary. Let me know if you prefer it is not
added.