Resolves #673: Add trace-level logging to inspect KeyValueUnsplitter
This allows the user to inspect what is going at a low level when doing a scan, but at a lower log level than most people should be running (as it is very chatty). I did enable trace level logging for that class for tests, but maybe that is a mistake (as it might be too much even for the Record Layer unit tests).