Fixes #682: Remove NullStandIn checking from the equality check for fields
This removes the `NullStandIn` checking in the `.equals` method from the `FieldKeyExpression`. It also removes it from that fields hash code methods, and it removes the `equivalentForSort` method from all key expressions, as it was only used to check for the `NullStandIn` anyway. It references #677 in a few places, which is the issue to move this into meta-data that is stored at the `RecordType` level, as having mixed values for different indexes does not really make sense.